[ticket] Ticket and review
Form, view and urls made for ticket and review.
We can now create and list from a loged session.

For now review is only created with an existing ticket.
Detail view only for ticket :
ticket detail visible for each review.

Name of application changed from Reviewer to
reviewer for convention.